flask-kafka/requirements.txt
flask
pyyaml
flasgger
marshmallow
confluent-kafka